Created a HitScan class
