+++ Assets/Code/Scripts/Bullets/HitScan.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74dc6a06948e9bb4b9287696588f9df8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
